requirements.txt
#Package                         Version
#------------------------------- -----------
altgraph                        0.17.4
annotated-types                 0.7.0
anyio                           4.8.0
async-generator                 1.10
attrs                           23.2.0
beautifulsoup4                  4.12.2
certifi                         2024.2.2
cffi                            1.16.0        
charset-normalizer              3.3.2
click                           8.2.1
colorama                        0.4.6
comtypes                        1.4.1
cryptography                    42.0.5        
distro                          1.9.0
grpcio                          1.71.0        
grpcio-tools                    1.71.0        
h11                             0.14.0        
httpcore                        1.0.7
httpx                           0.28.1        
idna                            3.7
jiter                           0.8.2
jsonpath-ng                     1.7.0
jsonschema                      4.24.0        
jsonschema-specifications       2025.4.1      
lxml                            5.2.1
natsort                         8.4.0
numpy                           2.3.0
openai                          1.64.0        
outcome                         1.3.0.post0   
overrides                       7.7.0
packaging                       25.0
pefile                          2023.2.7      
pip                             25.1.1        
ply                             3.11
protobuf                        5.29.3        
pycparser                       2.21
pydantic                        2.10.6        
pydantic_core                   2.27.2        
pydocstyle                      6.1.1
pyinstaller                     6.14.1        
pyinstaller-hooks-contrib       2025.5        
pyOpenSSL                       22.0.0        
PySocks                         1.7.1
python-docx                     0.8.11        
pywin32                         306
pywin32-ctypes                  0.2.3
pywinauto                       0.6.8
PyYAML                          6.0.1
referencing                     0.36.2        
requests                        2.31.0        
robotframework                  7.3
robotframework-assertion-engine 3.0.3
robotframework-browser          19.5.1        
robotframework-jsonlibrary      0.5
robotframework-pythonlibcore    4.4.1
robotframework-requests         0.9.5
robotframework-selenium2library 3.0.0
robotframework-seleniumlibrary  6.2.0
robotframework-yamllibrary      0.2.8
rpds-py                         0.26.0        
seedir                          0.5.0
selenium                        4.15.2        
setuptools                      80.9.0        
six                             1.16.0        
sniffio                         1.2.0
snowballstemmer                 2.1.0
sortedcontainers                2.4.0
soupsieve                       2.5
tqdm                            4.67.1        
trio                            0.20.0        
trio-websocket                  0.9.2
typing_extensions               4.12.2        
urllib3                         1.26.9        
wrapt                           1.17.2        
wsproto                         1.2.0